requirements.txt
playwright==1.49.0
pandas==2.2.3
streamlit==1.41.1
plotly==5.24.1
python-dotenv==1.0.1
requests
beautifulsoup4
supabase